Introduce convenient plugin for most basic native use case
Starting with this commit, I'm going to develop in parallel all native
plugins. The goal is to fit all the native plugin in the same track as
they should be. There will be differences such as Swift vs C++ vs C
compatibility, etc. However, they should all behave the same way. The
progress may be a bit slower than trying to figure out which plugin is
needed and asking existential questions on what is the best way forward.
Instead, the software model had a solution, so let's get back to the
same solution. There will be a lot of effort in building a composable
API that each plugin will use as an example. In the future, we may
retire those plugins, but I doubt this will happen as it give great
convenience to start a project.